+++ backend/matching_service/server.js
@@ -1,10 +1,10 @@
 const express = require("express");
-const app = express();
 const http = require("http");
 const { Server } = require("socket.io");
 const cors = require("cors");
-const { sequelize } = require("./db");
-const { roomModel, userModel } = require("./model");
+const { checkRoom } = require("./modelController");
+
+const app = express();
 app.use(cors());
 
 const server = http.createServer(app);
@@ -15,25 +15,6 @@ const io = new Server(server, {
   },
 });
 
-async function createRoom(socketID, difficulty) {
-  await roomModel.create({ socketID: socketID, difficulty: difficulty });
-}
-async function checkRoom(socket, difficulty) {
-  const rooms = await roomModel.findAll({
-    where: { difficulty: difficulty },
-  });
-  // console.log("matching room", rooms);
-  if (rooms.length > 0) {
-    console.log("room found");
-    socket.join(socket.id);
-    socket.emit("matched", rooms[0].socketID);
-    await roomModel.destroy({ where: { difficulty: difficulty } });
-  } else {
-    console.log("room not found");
-    // createUser(socket.id);
-    createRoom(socket.id, difficulty);
-  }
-}
 io.on("connection", (socket) => {
   socket.join(socket.id);
   console.log("a user connected:", socket.id);
